requirements.txt
Flask==3.0.2
Flask_MySQLdb==2.0.0
mysql_connector_repackaged==0.3.1
